+++ src/main/java/org/octopus/core/http/RedirectException.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.octopus.core.http;
+package org.octopus.core.http;
 
 public class RedirectException extends Exception {
     public RedirectException(String message) {
