+++ src/main/java/org/octopus/core/http/RedirectLimitException.java
@@ -22,21 +22,10 @@
  * SOFTWARE.
  */
 
-package com.octopus.core.http;
+package org.octopus.core.http;
 
-import java.net.URL;
-
-public class HTTPUtils {
-    /**
-     * Extract the filename from the URL(with extension)
-     *
-     * @param url URL that to be inspected
-     * @return filename
-     */
-    public static String extractFileNameFromURL(URL url) {
-        String filename;
-        String path = url.getPath();
-        filename = path.substring(path.lastIndexOf('/') + 1);
-        return filename;
+public class RedirectLimitException extends Exception {
+    public RedirectLimitException(int exceptions) {
+        super(System.out.printf("exceeded maximum %d redirects", exceptions).toString());
     }
 }
